+++ mdast_util_to_markdown/src/configure.rs
@@ -1,21 +1,21 @@
 #[allow(dead_code)]
 pub struct Options {
-    bullet: char,
-    bullet_other: char,
-    bullet_orderd: char,
-    emphasis: char,
-    fences: char,
-    list_item_indent: IndentOptions,
-    quote: char,
-    rule: char,
-    strong: char,
-    increment_list_marker: bool,
-    close_atx: bool,
-    resource_link: bool,
-    rule_spaces: bool,
-    set_text: bool,
-    tight_definitions: bool,
-    rule_repetition: u32,
+    pub bullet: char,
+    pub bullet_other: char,
+    pub bullet_orderd: char,
+    pub emphasis: char,
+    pub fences: char,
+    pub list_item_indent: IndentOptions,
+    pub quote: char,
+    pub rule: char,
+    pub strong: char,
+    pub increment_list_marker: bool,
+    pub close_atx: bool,
+    pub resource_link: bool,
+    pub rule_spaces: bool,
+    pub set_text: bool,
+    pub tight_definitions: bool,
+    pub rule_repetition: u32,
 }
 
 #[allow(dead_code)]
