+++ mdast_util_to_markdown/src/construct_name.rs
@@ -1,0 +1,31 @@
+#[derive(Clone, PartialEq)]
+#[allow(dead_code)]
+pub enum ConstructName {
+    Autolink,
+    Blockquote,
+    CodeIndented,
+    CodeFenced,
+    CodeFencedLangGraveAccent,
+    CodeFencedLangTilde,
+    CodeFencedMetaGraveAccent,
+    CodeFencedMetaTilde,
+    Definition,
+    DestinationLiteral,
+    DestinationRaw,
+    Emphasis,
+    HeadingAtx,
+    HeadingSetext,
+    Image,
+    ImageReference,
+    Label,
+    Link,
+    LinkReference,
+    List,
+    ListItem,
+    Paragraph,
+    Phrasing,
+    Reference,
+    Strong,
+    TitleApostrophe,
+    TitleQuote,
+}
