+++ mdast_util_to_markdown/src/util/check_strong.rs
@@ -1,0 +1,7 @@
+use alloc::string::String;
+
+use crate::state::State;
+
+pub fn check_strong(_state: &State) -> Result<char, String> {
+    Ok('*')
+}
